requirements.txt
Django
phonenumbers
django-phonenumber-field
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
django-environ
drf-spectacular


# [dev-packages]
isort
black
autoflake
flake8
pre-commit